updated ubuntu setup script.

+++ bash/ubuntu_server_setup.sh
@@ -48,11 +48,24 @@ sudo apt install -y \
 sudo systemctl disable systemd-networkd-wait-online.service
 sudo systemctl mask systemd-networkd-wait-online.service
 
-# remove snapd
-sudo snap remove lxd
-sudo snap remove core*
-sudo snap remove snapd
-sudo apt purge snapd -y
+# remove snaps.
+if [ -f /usr/bin/snap ]; then
+    snaps=$(snap list | awk 'NR>1 {print $1}')
+   # Need to remove all other snap packages before removing snapd.
+   if [ -z $snaps ]; then
+        for snap in $snaps; do
+            if[ $snap != 'snapd' ]; then
+                sudo snap remove $snap
+                echo "Removed snap package $snap"
+            fi
+        done
+        # now remove snapd.
+        sudo snap remove snapd
+        echo "Removed snap package snapd."
+    fi
+    # Blow that garbage away.
+    sudo apt purge snapd -y
+fi
 
 # finally update the system.
 sudo apt update && sudo apt dist-upgrade -y
